Add wasm wrapper
