+++ src/Makefile
@@ -60,6 +60,10 @@ blst: $(BLST_LIBRARY)
 c_kzg_4844.o: c_kzg_4844.c $(BLST_LIBRARY)
 	@$(CC) $(CFLAGS) -c $<
 
+wasm.o: wasm.c c_kzg_4844.o 
+	@$(CC) $(CFLAGS) -c $<
+
+
 test_c_kzg_4844: CFLAGS += -O0
 test_c_kzg_4844: test_c_kzg_4844.c c_kzg_4844.c $(BLST_LIBRARY)
 	@$(CC) $(CFLAGS) -o $@ $< $(LIBS)
