+++ src/c_kzg_4844.h
@@ -180,8 +180,6 @@ C_KZG_RET verify_blob_kzg_proof_batch(
     const KZGSettings *s
 );
 
-C_KZG_RET load_trusted_setup_file_from_wasm();
-
 #ifdef __cplusplus
 }
 #endif
